blobstore: generalize peapod/bbcz benchmarks
BBCZ is deprecated, so it's not added into generic Put test, but the test
itself is transferred.
